Do 55_Hustle ability

+++ src/pokemon/model/Attack.java
@@ -17,8 +17,8 @@ public class Attack {
 	private int power;
 	private int initialPower;
 	private int pp;
-	private int precision;
-	private int initialPrecision;
+	private float precision;
+	private float initialPrecision;
 	private String effect;
 	private PokemonType strTypeToPkType;
 	private float effectivenessAgainstPkFacing;
@@ -59,7 +59,7 @@ public class Attack {
 		this.reduceStats = false;
 	}
 
-	public Attack(int id, String name, String type, int power, int pp, int precision, String effect) {
+	public Attack(int id, String name, String type, int power, int pp, float precision, String effect) {
 		super();
 		this.id = id;
 		this.name = name;
@@ -81,7 +81,7 @@ public class Attack {
 		this.hasSecondaryEffect = false;
 		this.reduceStats = false;
 	}
-	
+
 	public Attack(Attack attack) {
 		super();
 		this.id = attack.id;
@@ -158,11 +158,11 @@ public class Attack {
 		this.pp = pp;
 	}
 
-	public int getPrecision() {
+	public float getPrecision() {
 		return precision;
 	}
 
-	public void setPrecision(int precision) {
+	public void setPrecision(float precision) {
 		this.precision = precision;
 	}
 
@@ -243,11 +243,11 @@ public class Attack {
 		this.initialPower = initialPower;
 	}
 
-	public int getInitialPrecision() {
+	public float getInitialPrecision() {
 		return initialPrecision;
 	}
 
-	public void setInitialPrecision(int initialPrecision) {
+	public void setInitialPrecision(float initialPrecision) {
 		this.initialPrecision = initialPrecision;
 	}
 
@@ -286,7 +286,7 @@ public class Attack {
 	public List<SecondaryEffect> getSecondaryEffects() {
 		return secondaryEffects;
 	}
-	
+
 	public void setSecondaryEffectsNull() {
 		this.secondaryEffects = null;
 	}
